% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prof_log_lik}
\alias{prof_log_lik}
\title{Calculate (profile) log-likelihood for Sigma_c, i.e. the multivariate normal likelihood 
proportional to: -log(determinant(Sigma)) - trace{t(E) solve(Sigma) E},
 where Sigma =  Sigma_c t(Sigma_c).}
\usage{
prof_log_lik(Sigma_c, E)
}
\arguments{
\item{Sigma_c}{An rc x rc lower triangular Cholesky factor to evaluate the likelihood at}

\item{E}{An rc x n_obs matrix of residual vectors, where n_obs is the number of such vectors}
}
\value{
The likelihood value evaluated at the arguments supplied, *including constants*
}
\description{
Calculate (profile) log-likelihood for Sigma_c, i.e. the multivariate normal likelihood 
proportional to: -log(determinant(Sigma)) - trace{t(E) solve(Sigma) E},
 where Sigma =  Sigma_c t(Sigma_c).
}
