% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sepcor}
\alias{sepcor}
\title{Estimate the covariance matrix Sigma = W k(U, V) W using the multivariate
normal likelihood, where k(U, V) means the Kronecker product of U and V.}
\usage{
sepcor(E, n_rows, sepcov = FALSE, tol = 1e-16, maxiter = 1000,
  verbose = FALSE)
}
\arguments{
\item{E}{Matrix of dimension (n_rows n_cols) x n_obs, each column is a residual vector}

\item{n_rows}{Number of "rows" of the inverse vectorized columns of E}

\item{sepcov}{If TRUE, assume separable covariance matrix}

\item{tol}{Algorithm terminates when an iteration increases the log-likelihood less than tol}

\item{maxiter}{The maximum number of iterations the algorithm runs if not converging before}

\item{verbose}{Print additional info about iterates if TRUE}
}
\value{
Final iterates, log-likelihood evaluated at these iterates, iterations,
and convergence info
}
\description{
Estimate the covariance matrix Sigma = W k(U, V) W using the multivariate
normal likelihood, where k(U, V) means the Kronecker product of U and V.
}
